Cleanup API and add docs
